+++ lib/main.dart
@@ -5,6 +5,9 @@ import 'package:meta_app/presentation/navigation/app_router.gr.dart';
 import 'package:meta_app/presentation/redux/locale/actions/init_locale_action.dart';
 import 'package:meta_app/presentation/redux/store.dart';
 import 'package:meta_app/presentation/redux/theme/actions/init_theme_action.dart';
+import 'package:provider/provider.dart';
+
+import 'presentation/pages/client_profile_page/client_tabs_manager.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -15,5 +18,12 @@ void main() async {
   final appRouter = AppRouter();
   store.dispatch(InitThemeAction());
   store.dispatch(InitLocaleAction());
-  runApp(Application(store: store, appRouter: appRouter));
+  runApp(
+    MultiProvider(
+      providers: [
+        ChangeNotifierProvider(create: (context) => ClientTabsManager()),
+      ],
+      child: Application(store: store, appRouter: appRouter),
+    ),
+  );
 }
