+++ lib/presentation/navigation/app_router.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:meta_app/presentation/pages/blog_page.dart/blog_page.dart';
-import 'package:meta_app/presentation/pages/client_dashboard_page/client_dashboard_page.dart';
+import 'package:meta_app/presentation/pages/client_profile_page/client_profile_page.dart';
 import 'package:meta_app/presentation/pages/directions_page/directions_page.dart';
 import 'package:meta_app/presentation/pages/forgot_password_page.dart';
 import 'package:meta_app/presentation/pages/home_page/home_page.dart';
@@ -21,8 +21,8 @@ import 'package:meta_app/presentation/pages/registration_page.dart';
     AutoRoute(page: BlogPage, path: "/blog"),
     AutoRoute(page: DirectionsPage, path: "/directions"),
     AutoRoute(
-      page: ClientDashboardPage,
-      path: "/client-dashboard",
+      page: ClientProfilePage,
+      path: "/client-profile",
       initial: true,
     ),
   ],
