+++ lib/presentation/pages/client_profile_page/client_tabs_manager.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class ClientTabsManager with ChangeNotifier {
+  int _currentIndex = 0;
+
+  int get currentIndex => _currentIndex;
+
+  void changeIndex(int index) {
+    _currentIndex = index;
+    notifyListeners();
+  }
+}
