+++ lib/presentation/pages/client_profile_page/sections/bots_tab.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class BotsTab extends StatelessWidget {
+  const BotsTab({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      color: const Color(0xFFE5E5E5),
+      child: const Center(
+        child: Text('Bots'),
+      ),
+    );
+  }
+}
