+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   flutter_bloc: ^8.1.1
   async_redux: ^17.0.0
   rxdart: ^0.27.5
+  provider: ^6.0.5
 
   # Api
   retrofit: ^3.0.1+1
